add connection

+++ app/Controllers/Http/NodeConnectionsController.ts
@@ -5,34 +5,53 @@ export default class NodeConnectionsController {
 
   public async show ({ params }: HttpContextContract) {
     const id = parseInt(params.id)
+    console.log(id)
     return prisma.$queryRaw`
     select n.*, t.name as nodeTypeName, t.color as nodeTypeColor from Node n
     join NodeConnection c on n.id = c.connectedToId
     join NodeType t on n.typeId = t.id
-    where c.connectedFromId = ${id}`
+    where c.connectedFromId = ${id}
+
+    union
+
+    select n.*, t.name as nodeTypeName, t.color as nodeTypeColor from Node n
+    join NodeConnection c on n.id = c.connectedFromId
+    join NodeType t on n.typeId = t.id
+    where (c.connectedToId = ${id})
+    `
   }
 
-  public async store ({ params, response }: HttpContextContract) {
-    const from = parseInt(params.from)
-    const to = parseInt(params.to)
+  public async store ({ request, response }: HttpContextContract) {
+    const body = request.only(['from', 'to'])
+    const from = parseInt(body.from)
+    const to = parseInt(body.to)
 
     const result = await prisma.nodeConnection.findFirst({
       where: {
         OR: [
-          {
-            connectedFromId: from,
-            connectedToId: to
-          },
-          {
-            connectedFromId: to,
-            connectedToId: from
-          }
+          {AND: [
+            {
+              connectedFromId: from,
+            },
+            {
+              connectedToId: to
+            }
+          ]},
+          {AND: [
+            {
+              connectedFromId: to,
+            },
+            {
+              connectedToId: from
+            }
+          ]}
+
         ]
       }
     })
 
     if(result) {
-      return response.badRequest('Connection already exists')
+      return response.send({message: 'Connection already exists', status: 409})
     }
 
     await prisma.nodeConnection.create({
@@ -41,14 +60,46 @@ export default class NodeConnectionsController {
         connectedToId: to
       }
     })
+    return response.json({message: 'Connection created', status: 201})
   }
 
-  public async destroy ({ params }: HttpContextContract) {
+  public async destroy ({ params, response }: HttpContextContract) {
     const id = parseInt(params.id)
-    return await prisma.nodeConnection.delete({
+    const connections = await prisma.nodeConnection.findUnique({
       where: {
         id: id
       }
     })
+    if(connections) {
+      var from = connections.connectedFromId
+      var to = connections.connectedToId
+
+      const connectionsToDestroy = await prisma.nodeConnection.findMany({
+        where: {
+          OR: [
+            {
+              connectedFromId: from,
+              connectedToId: to
+            },
+            {
+              connectedFromId: to,
+              connectedToId: from
+            }
+          ]
+        }
+      })
+
+      connectionsToDestroy.forEach(async (connection: any) => {
+        await prisma.nodeConnection.delete({
+          where: {
+            id: connection.id
+          }
+        })
+      })
+      return response.ok('Connection deleted')
+    }
+    else {
+      return response.badRequest('Connection does not exist')
+    }
   }
 }
